Remove __test feature
